+++ commands/emotereply.js
@@ -16,6 +16,8 @@ module.exports = {
         if (response === "") { return; }
         message.channel.send(response);
 
-        setTimeout(() => message.delete().catch(console.log), 2000);
+        if (message.channel.type === "text") {
+            setTimeout(() => message.delete().catch(console.log), 2000);
+        }
     }
 };
