+++ commands/reaction.js
@@ -22,6 +22,8 @@ module.exports = {
         
         messages[0].react(client.emotes.get(emote)).catch(console.log);
 
-        setTimeout(() => message.delete().catch(console.log), 2000);
+        if (message.channel.type === "text") {
+            setTimeout(() => message.delete().catch(console.log), 2000);
+        }
     }
 };
